Moved local physical tests to the right test
